+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## New
+- Added functions to get a timestamp from a verified signature, to the crypto module and to the KeyHolder's context:
+    `getVerifiedTimestampOfText` and `getVerifiedTimestampOfData`
+
 ### Changes
 - Added validation for Client Version header (x-pm-appversion).